--- src/assets/ProductsCard.tsx
@@ -1,28 +1,0 @@
-const productCards = [
-    {
-        id: 1,
-        name: "Product Card № 1",
-    },
-    {
-        id: 2,
-        name: "Product Card № 2",
-    },
-    {
-        id: 3,
-        name: "Product Card № 3",
-    },
-    {
-        id: 4,
-        name: "Product Card № 4",
-    },
-    {
-        id: 5,
-        name: "Product Card № 5",
-    },
-    {
-        id: 6,
-        name: "Product Card № 6",
-    },
-];
-
-export default productCards;
