--- src/assets/customers.tsx
@@ -1,77 +1,0 @@
-
-const customers = [
-    {
-        id: 1,
-        fName: "Aurora",
-        lName: "West",
-        email: "",
-        password: "",
-        companyName: "",
-        website: "",
-        cartProduct: [],
-        card: "",
-        notes: [],
-    },
-    {
-        id: 2,
-        fName: "Lucas",
-        lName: "Brooks",
-        email: "",
-        password: "",
-        companyName: "",
-        website: "",
-        cartProduct: [],
-        card: "",
-        notes: [],
-    },
-    {
-        id: 3,
-        fName: "Seraphina",
-        lName: "Patel",
-        email: "",
-        password: "",
-        companyName: "",
-        website: "",
-        cartProduct: [],
-        card: "",
-        notes: [],
-    },
-    {
-        id: 4,
-        fName: "Jasper",
-        lName: "Nguyen",
-        email: "",
-        password: "",
-        companyName: "",
-        website: "",
-        cartProduct: [],
-        card: "",
-        notes: [],
-    },
-    {
-        id: 5,
-        fName: "Eliza",
-        lName: "Monroe",
-        email: "",
-        password: "",
-        companyName: "",
-        website: "",
-        cartProduct: [],
-        card: "",
-        notes: [],
-    },
-    {
-        id: 6,
-        fName: "Xavier",
-        lName: "Chavez",
-        email: "",
-        password: "",
-        companyName: "",
-        website: "",
-        cartProduct: [],
-        card: "",
-        notes: ["left a good tips"],
-    },
-];
-
-export default customers;
