--- src/assets/stripeTerminals.tsx
@@ -1,13 +1,0 @@
-const stripeTerminals = [
-    {
-        id: 1,
-        name: "Terminal № 1",
-    },
-    {
-        id: 2,
-        name: "Terminal № 2",
-    },
-];
-
-export default stripeTerminals;
-
