+++ src/modules/customer-order-notes/template/CustomerOrderNote.tsx
@@ -1,12 +1,10 @@
 import { Link, useNavigate } from "react-router-dom";
 import { useState, useEffect } from "react";
 import { useQueryClient, useMutation, useQuery } from "@tanstack/react-query";
-import Note from "../components/Notes";
-import { Input } from "@medusajs/ui";
-import { Button } from "@medusajs/ui";
+import Note from "../../../components/Notes";
 import { ArrowDownLeftMini } from "@medusajs/icons";
-
-const DraftOrderNote = ({ setClient, client }: { setDraftOrder: (newOrder: boolean) => void; newOrder: boolean; client: any, setClient: (client: any) => void }) => {
+import { Button, Input } from "@medusajs/ui";
+const OrderNote = ({ setClient, client, setEnable }: { setDraftOrder: (newOrder: boolean) => void; newOrder: boolean; setEnable: any, client: any, setClient: (client: any) => void }) => {
     const navigate = useNavigate();
     const [note, setNote] = useState("");
     const queryClient = useQueryClient();
@@ -15,30 +13,28 @@ const DraftOrderNote = ({ setClient, client }: { setDraftOrder: (newOrder: boole
         const storedClient = localStorage.getItem("client");
         if (storedClient) {
             setClient(JSON.parse(storedClient));
+            setEnable(false);
         }
-    }, [setClient]);
+    }, [setClient, setEnable]);
 
     const { isLoading: draftsIsLoading, isError: draftsIsError, data: draftsData, error: draftsError } = useQuery(
-        ["draftNotes", client?.cart.id],
+        ["notes", client?.id],
         async () => {
-            if (!client?.cart.id) throw new Error("Client ID is not defined");
-            const response = await fetch(`http://localhost:9000/admin/custom/cartOrderNote/${client.cart.id}`, {
+            if (!client?.id) throw new Error("Client ID is not defined");
+            const response = await fetch(`http://localhost:9000/admin/custom/customer/${client.id}`, {
                 method: "GET",
                 credentials: 'include',
             });
-
             if (!response.ok) {
                 throw new Error("Failed to fetch data");
             }
             return response.json();
-        },
-        {
-            enabled: !!client?.cart?.id, // Runs only when client.id is defined
         }
     );
+
     const submitNoteMutation = useMutation({
         mutationFn: async (bodyNote: { title: string }) => {
-            const response = await fetch(`http://localhost:9000/admin/custom/cartOrderNote/${client.cart.id}`, {
+            const response = await fetch(`http://localhost:9000/admin/custom/customer/${client.id}`, {
                 method: "PUT",
                 credentials: "include",
                 body: JSON.stringify({ bodyNote }),
@@ -53,14 +49,15 @@ const DraftOrderNote = ({ setClient, client }: { setDraftOrder: (newOrder: boole
             }
             return response.json();
         },
-        onSuccess: () => {
+        onSuccess: (data) => {
+            queryClient.invalidateQueries(["notes", client.id]);
             setNote("");
-            queryClient.invalidateQueries(["draftNotes", client.cart.id]);
-        }
+            console.log("Note added:", data);
+        },
     });
     const deleteNotes = useMutation({
         mutationFn: async (noteId) => {
-            const response = await fetch(`http://localhost:9000/admin/custom/cartOrderNote/${client.cart.id}`, {
+            const response = await fetch(`http://localhost:9000/admin/custom/customer/${client.id}`, {
                 method: "DELETE",
                 body: JSON.stringify({ noteId }),
                 headers: {
@@ -76,7 +73,7 @@ const DraftOrderNote = ({ setClient, client }: { setDraftOrder: (newOrder: boole
             return response.json();
         },
         onSuccess: ((data) => {
-            queryClient.invalidateQueries({ queryKey: ["draftNotes", client.cart.id] });
+            queryClient.invalidateQueries({ queryKey: ["notes", client.id] });
             console.log("Note Deleted:", data);
         }),
         onError: (error) => {
@@ -101,6 +98,7 @@ const DraftOrderNote = ({ setClient, client }: { setDraftOrder: (newOrder: boole
     const deleteNote = (noteId: any) => {
         deleteNotes.mutate(noteId);
     }
+
     return (
         <>
             <nav className="w-full px-[25px] h-[80px] flex flex-row flex-nowrap justify-between items-center border">
@@ -109,11 +107,11 @@ const DraftOrderNote = ({ setClient, client }: { setDraftOrder: (newOrder: boole
                 </Link>
             </nav>
             <header className="p-[15px] self-start border-b border-b-[rgba(206,206,206,1)]">
-                <p className="text-[15px] text-gray-400">Draft Order Notes</p>
+                <p className="text-[15px] text-gray-400">Customer Notes</p>
             </header>
             <main className="flex flex-col flex-nowrap justify-center items-center gap-[20px] my-[20px]">
                 <div className="flex flex-col flex-nowrap justify-center items-center gap-[10px]">
-                    <Input type="text" placeholder="Order Notes" onChange={handleChange} value={note} className="h-[50px] w-[400px]" />
+                    <Input type="text" placeholder="New Note..." onChange={handleChange} value={note} className="h-[50px] w-[400px]" />
                     <Button type="submit" className="btn-order-note" onClick={() => handleSubmit({ title: note })}>Submit</Button>
                 </div>
                 <div className="flex flex-row flex-wrap justify-start items-center gap-[15px] self-start mx-[20px]">
@@ -126,4 +124,4 @@ const DraftOrderNote = ({ setClient, client }: { setDraftOrder: (newOrder: boole
     );
 };
 
-export default DraftOrderNote;
+export default OrderNote;
