+++ src/modules/terminals/component/StripeTerminals.tsx
