+++ src/modules/type-of-customers/template/TypeOfCustomer.tsx
